pyproject.toml
[tool.poetry]
name = "mongofitness"
version = "0.1.0"
description = "upload fitness stats from Strava and Polar to a MongoDB"
authors = ["Anders Retzner <anders.retzner@gmail.com>"]
readme = "README.md"
packages = [{include = "mdbfit"}]

[tool.poetry.dependencies]
python = "^3.11"
requests-oauthlib = "^1.3.1"
load-dotenv = "^0.1.0"
appdirs = "^1.4.4"
pymongo = "^4.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
